Extract useBubble custom hook
